% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{eu_num}
\alias{eu_num}
\title{Create an EU-formatted numeric vector
This constructor wraps a numeric vector with the `eu_num` S3 class, allowing it to print using European number formatting (comma as decimal mark, dot as thousands separator).}
\usage{
eu_num(x, digits = 2, currency = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{digits}{Integer vector specifying decimal digits.
Can be of length 1 or the same length as `x`.}

\item{currency}{Logical; if `TRUE` append `" €"` to formatted output.}
}
\value{
An object of class `eu_num` (and `"numeric"` underneath).
}
\description{
Create an EU-formatted numeric vector
This constructor wraps a numeric vector with the `eu_num` S3 class, allowing it to print using European number formatting (comma as decimal mark, dot as thousands separator).
}
\examples{
eu_num(pi)
}
